+++ test/platform.rb
@@ -1,0 +1,7 @@
+class Test
+  class Platform_Test < MTest::Unit::TestCaseWithAnalytics
+    def test_released?
+      assert_true released?
+    end
+  end
+end